error wrapping, refactor
